services/face_reid/requirements.txt
# Requirements placeholder

fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.1
torchvision==0.16.1
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
insightface==0.7.3
redis==5.0.1
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.2
aiohttp==3.9.1
asyncio==3.4.3
